Add React installation instructions to web docs
This might seem obvious, but it really isn't for a newcommer. Note that this is mentioned in README.md in source code.

I skipped mention about only needing to install react when you create App as opposed to library, because I guess 99% of users are creating the app and those who create a library know what they are doing.